refactor(core): Move files to agent/app
