+++ core/agent/app/__init__.py
@@ -1,0 +1,2 @@
+"""Core agent app-level modules."""
+
